% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lhs.R
\name{get_lhs}
\alias{get_lhs}
\title{Get left hand side of a formula}
\usage{
get_lhs(form)
}
\arguments{
\item{form}{Either a string in the form \code{'y ~ ...'} or an object of \code{formula} class}
}
\value{
A string with the name of the left hand side variable in the formula
}
\description{
Returns the dependent variable in a formula given by a string or a \code{formula}
}
