% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from a formula}
\usage{
get_parameters(form, var_names)
}
\arguments{
\item{form}{Either a string in the form \code{'y ~ ...'} or an object of \code{formula} class}

\item{var_names}{A string array with the column names of the data.frame containing the data to be fitted.}
}
\value{
A list containing the names of the linear and the nonlinear parameters of the formula.
}
\description{
Returns the linear and nonlinear parameters of a \code{formula}
}
