% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nlstac.R
\name{is.nlstac}
\alias{is.nlstac}
\title{Is nlsTAC class check}
\usage{
is.nlstac(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a tac object (that is, has "tac"
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks wether an R object is of tac class or not.
}
