% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nlstac.R
\name{print.nlstac}
\alias{print.nlstac}
\title{Print a nlstac Model}
\usage{
\method{print}{nlstac}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"nlstac"} obtained by the \code{nls_tac} function.}

\item{digits}{a positive integer indicating how many significant digits are to be shown.}

\item{...}{Ignored, for compatibility issues.}
}
\description{
Standard method for overriding the print.list method for nlstac model fit.
}
\author{
\strong{Mariano Rodríguez-Arias} (\email{arias@unex.es}).
\emph{Deptartment of Mathematics}

\strong{Juan Antonio Fernández Torvisco} (\email{jfernandck@alumnos.unex.es}).
\emph{Department of Mathematics}

University of Extremadura (Spain)

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
