% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.nlstac.R
\name{print.summary.nlstac}
\alias{print.summary.nlstac}
\title{Prints the summary a summary.nlstac object.}
\usage{
\method{print}{summary.nlstac}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"nlstac"} obtained by the \code{fit_tac} function.}

\item{digits}{Number of significant digits to be shown (defaults to 3).}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{Ignored, for compatibility issues.}
}
\description{
Internal function for printing the summary of a nlstac.
}
