% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.nlstac.R
\name{residuals.nlstac}
\alias{residuals.nlstac}
\title{Extract Model Residuals for a nsltac fit model}
\usage{
\method{residuals}{nlstac}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nlstac"} obtained by the \code{nls_tac} function.}

\item{...}{Ignored, for compatibility issues.}
}
\value{
A vector with the residual values.
}
\description{
Returns the model residuals of the fit.
}
\author{
\strong{Mariano Rodríguez-Arias} (\email{arias@unex.es}).
\emph{Deptartment of Mathematics}

\strong{Juan Antonio Fernández Torvisco} (\email{jfernandck@alumnos.unex.es}).
\emph{Department of Mathematics}

University of Extremadura (Spain)

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
