% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nlstac.R
\name{summary.nlstac}
\alias{summary.nlstac}
\title{Summary a nls tac fit.}
\usage{
\method{summary}{nlstac}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nlstac"} obtained by the \code{fit_tac} function.}

\item{...}{Ignored, for compatibility issues.}
}
\value{
Returns, via the \code{print.nlstac} function the following items:
    - Formula: The formula fitted to the data
    - Parameters: The value of the estimated parameters (Estimated) together
                  with their standard errors (Std. Error), and their statistical
                  significance (t value, Pr(>|t|), signif. stars)
    - SSR and df.
    - Convergence information: N. of iterations and the tolerance achieved.
}
\description{
Gives the fitted coefficients and the convergence information of the fit.
}
