% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.nlstac.R
\name{vcov.nlstac}
\alias{vcov.nlstac}
\title{Calculate Variance-Covariance Matrix for a nlstac Fitted Model Object}
\usage{
\method{vcov}{nlstac}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"nlstac"} obtained by the \code{nls_tac} function.}

\item{...}{Ignored, for compatibility issues.}
}
\value{
A matrix of the estimated covariances between the parameter estimates.
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted model object. 
The “main” parameters of model correspond to those returned by coef,
}
\author{
\strong{Mariano Rodríguez-Arias} (\email{arias@unex.es}).
\emph{Deptartment of Mathematics}

\strong{Juan Antonio Fernández Torvisco} (\email{jfernandck@alumnos.unex.es}).
\emph{Department of Mathematics}

University of Extremadura (Spain)

\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
